refactor: 패키지 변경
- 설정과 관련된 정보만 config.client 에 남아있도록
