+++ backend/src/main/java/reviewme/auth/infrastructure/dto/response/GitHubAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package reviewme.config.client.dto.response;
+package reviewme.auth.infrastructure.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
